pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler==3.6.3",
    "flask>=3.1.0",
    "flask-wtf>=1.2.2",
    "pytelegrambotapi>=4.26.0",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
]
